requirements.txt
streamlit
pandas
geopandas
folium
streamlit-folium
shapely
fiona
pyproj
numpy
requests
openpyxl
PyYAML
bcrypt
google-cloud-storage
streamlit-authenticator>=0.2.3
